+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+# Version 1.8.1
+* Popup enhancements:
+  * Allow filter to match URLs as well as link text
+  * Separate hiding duplicates from deduplicating when opening tabs/windows
 # Version 1.8.0
 * Port to Typescript
   - this brings more assurance of correctness.
 * New OSL options
   - `deduplicate`: when true, only open one tab for identical links in the selection.
   - `focus`: when true, make the opened tab active or give the opened window focus.  
-    *Note*: New windows may be drawn on top without being given focus, which can be confusing.=
+    *Note*: New windows may be drawn on top without being given focus, which can be confusing.
 * Add persistent settings
   - no longer forced to re-select options in the popup.
   - can customize affect context menu clicks.